Fetch new commit from upstream
Well I've decided to do a "refresh" because I am kinda overwhelmed on
what to do if the upstream receives new updates. This is also a good
time to manually modify and understand the CSS. Any future upstream
updates will be carefully merged, so we can discard any unneeded changes.

As the first round of refresh, we will just be fetching the new commits
as is, with no modifications. Yes it will overwrite the current CSS but
that's what git for ;) Anyways I've did the necessary back-ups in case
anything breaks.

Upstream changes:
MrOtherGuy/firefox-csshacks@233d758
hide_tabs_toolbar: make the style work better with titlebar and menubar