Remove commented CSS
And for the 3rd round of refresh, I will actually remove
the unnescessary CSS code.

Reasons the CSS code is removed include:
- I don't use that feature (compact mode)
- Doesn't appear to do anything
- Changes for other OS

I have a principle to only change what's necessary. ;)
If the line of CSS doesn't seem to do anything, and isn't
required to function, then I will by all means remove it.
I don't want my userChrome.css to break any future changes
when Firefox updates.

As for why to make another commit just to remove the code,
well this is what git is for, version control. I can see previous
commit to see the code that has been deleted. Who knows, one of the
CSS that I deemed "it doesn't do anything" might be needed later on.