Add drop shadow when sidebar is hovered
The sidebar felt out of place when I got rid of the CSS for divider in
4190f34, so I wanted to re-add it back. Who knew how much difference a
1px line could make :P Then I thought about an idea - drop shadow. As a
previous Edge user, that's one of the thing I felt missing with vertical
tabs in Firefox. Without it, it just doesn't feel like it's part of the
UI. Soo here it is!

From https://developer.mozilla.org/en-US/docs/Web/CSS/box-shadow

/* offset-x | offset-y | blur-radius | spread-radius | color */
box-shadow: 0px 6px 12px 0 rgba(0, 0, 0, 0.5)

6px in offset-y is needed if not there's shadow above the sidebar D: