Add configuration for nightTab to customise newtab
nightTab_base.css is well, the base for any of my future themes,
or more appropriately to be used as a comfy fallback.

I don't even know what's this theme called, but I just decided to
make it coherent with the background, using a combination of
lavender and purple.

Don't mind the bookmarks and be sure to change it to your own. ;)
This is just a backup for nightTab after all.