Add "nightTab_nord.json"
Nord version because... it's fitting for Christmas! Can't relate with
winter because it doesn't snow here, heh.

Anyways, just a new theme if I ever get bored with the base theme.
Here, I experimented a little with the layout. The header area is on the
left, and the bookmarks are on the right.

I also used • as the date separator, which looks kinda cool
(no pun intended) and added a snowflake emoji to the new tab title.

and that's it, yeah.