Only apply pointer fix to the tabs panel
The pointer fix is applied globally. Since the bookmark panel doesn't
have any padding at the top, the pointer is not aligned correctly.

`:not(.Sidebar[data-panel-type="bookmarks"])` checks if the current
panel is the bookmarks panel, if it isn't, apply the pointer fix.

For some reason `:not(.panel-btn[data-type="bookmark"][data-active="true"])`
didn't work, but I guess the above solution will do and it looks cleaner
as well.