Fix Tab Session Manager's category filter dropdown
Kinda niche, but it did bother me a bit. Since it's caused by my CSS,
that means it should be fixable by userChrome.css as well.

This fixes the padding on both sides of the dropdown, so it is consistent
with the top and bottom padding. I also decreased the space between each
option inside the dropdown, which makes it take about the same as before.

Unfortunately, there's still other side effects of proton_context_menu.css
that I've recently discovered:

- Any context menu in the dev panels are big and bulky. Looking through
  the Browser Toolbox, they don't seem to be grouped under any class.
  I can only see each ID for the individual context menu entries. This
  makes fixing it pretty hard.

- Context menu for page actions. Not used very often, but might as well
  make it complete. This seems simple enough at least.