Add sidebery-data.json
Just a backup of Sidebery's settings. As for why, because I'm going to
switch to Tree Style Tabs soon.

Sidebery has gotten increasingly unstable due to it not receiving any
updates since 4th June (that's 6 months)! You can just go its issues
section (https://github.com/mbnuqw/sidebery/issues) and see the amount
of pressing issues, many of them which are duplicates.

Despite this, I've been coping with its issues:
- Panels being plain broken? Just don't use it.
- Restore previous session not restoring tree hierachy? Tab Session
  Manager fixes that.
- Can't drag panel to create new window?
  Right-click tab -> Move to new window
and other numerous issues...

But the last straw is Sidebery's tabs can no longer be moved around on
Nightly. Now a very kind Mozilla engineer sent a pull-request
(mbnuqw/sidebery/issues/539) that fixes this major issue, but the main
problem still exists: the author isn't active. There's this fork which
tries to fix some of its core issues (https://github.com/pintassilgo/sidebery/),
but where's the compiled and signed extension?

Of course, I've merged both the fork's changes and the pull request,
then compiled the .xpi myself, but the extension is unsigned, so it can
only work on Nightly. So we are back to square one, what's the point? If
Sidebery breaks on Nightly, it's going to break on Stable, sooner or
later. Until the Sidebery's author is back from hiatus, or there's an
active fork to supply a signed extension, I will have to sadly leave
Sidebery.

Now, there's a reason I decided to back up its data: I'm interested in
using it, and will always prefer it, and when its core issues start
getting fixed, I will use it again~