Add dashed lines to child tabs
Yes, it is directly inspired from FlyingFox, and turns out it's only 1
line of CSS responsible for it! Here, I tried to align the dashed lines
with the twister, unfortunately, that will also move the tabs further to
the right, so there's a big gap between the dashed lines and the tabs.

When the sidebar is collapsed, the margin is reset back to 0px, so you
can still see the dashed line and the tabs even after its collapsed.