Add thin line beside tabs to indicate sub-tree level
Originally inspired from FlyingFox's TST CSS, now on Sidebery! Well, it
is not the exact same style because I try to use a unique line style
based on the which vertical tabs extension is used. Tree Style Tabs uses
the dashed line style, while Sidebery uses the solid thin line style. I
styled it to be a thin solid line because it's that way in the bookmarks
panel too.

Sidebery has an option for "Show marks to indicate tabs sub-tree levels",
but unfortunately there isn't any way to customize the mark symbol.

I try to replicate the style from treestyletabs.css in Sidebery:

.Tab:not([data-lvl="0"]) {
    border-left: solid 1px rgb(79, 79, 79)
}

which didn't work, because I styled the tabs to have rounded corner, so
the border lines would be rounded as well. Under every .Tab there is a
.lvl-wrapper class, so I can hook on that instead, and voila! The only
thing is the solid line will get disconnected if there's a child tab, but
imo this indicates the sub-tree level better than TST's style.