Remove comment about padding between tabs
It's possible, but somewhat hacky. You can see how it's done here:
https://www.reddit.com/r/FirefoxCSS/comments/rmi8dg/yet_another_sidebery_setup/

It involves removing the tabs default background, and instead apply the
background to .lvl-wrapper. Since .lvl-wrapper is slightly smaller than
.Tab, it will appear to have some space in between the tabs.

However, changing the padding in between the tabs isn't going to be that
straightforward, because we are piggybagging on the .lvl-wrapper element
instead of the .Tab element, which is what you are *supposed* to be using.

A nice side effect of this is that the tab highlight now starts from its
tree level (starting from the thin line), instead of being the same width
regardless of the tree level. This matches Tree Style Tabs's behaviour.

Whether I'm going to use this... I'm leaning towards "no", due to how
hacky it is, and I kinda like how close the tabs are, it's like they are
cuddling each other! Either way, adding padding between tabs is indeed
possible in Sidebery, and it's not exclusive to Tree Style Tabs only.