Make Sidebery follow Firefox dark theme
`sidebery-data.json` is updated as well to change the `--bg` variable
globally. The reason we can't specify it under the Sidebar CSS, is well,
because it only affects the sidebar. To apply the background color
universally, including the group page and the settings page, it has to be
specified manually under Styles Editor.