Add overscrolling after sidebar is full
This is a built-in feature to assist mouse-clicking on the tabs panel
when the panel is full. I couldn't find what's causing the regression,
so I just suck it up and moved on. Finally, when I was refactoring the
code earlier, I've identify the issue - my CSS was overwriting the
bottom padding from Sidebery's default.

64px isn't just any arbitrary value, it's the default value set when
you inspect Sidebery's panel in about:debugging. Of course, feel free
to increase the padding depending on how far you want to scroll after
reaching the end of tabs panel.