Preserve overscrolling when sidebar is collapsed
Previously, the sidebar will get rid of the additional overscroll space
if it's collapsed. This is especially annoying if you intended to leave
some space at the bottom (presumably so you can open a new tab later),
and the overscroll space will be reset every time the sidebar is
collapsed and reopened. The reason is because `padding: 2px` will
overwrite `padding-bottom: 64px` when `@media screen` kicks in.