Add sidebar divider
I initially hated the thin border below the URL bar/bookmarks toolbar,
but when I tried to remove it, the UI looked very weird. I also noticed
that literally every other browsers have a thin border around the
webpage. Thus, I added a divider to the sidebar to visually separate it
from the webpage, similar to the toolbar.