Update to Sidebery v5
I'll still keep the older `sidebery.css` since v5 is still in beta.
However, `sidebery-data.json`, which is my backup file, will continue to
be updated, so compatibility is not guaranteed.

"Displaying the active panel button when sidebar is collapsed" is harder
to pull off now. One, another class is blocking the element we're
interested in. Two, the panel button element disappears when collapsed,
so I have to override the `opacity` and `transform` properties. :shrug: