Improve nested tabs styling
This makes it easier to identify the tab level when the sidebar is
collapsed. I've stopped at [data-lvl="6"] because that's as deep as I
would use it (hopefully). Even so, at [data-lvl="5"] and [data-lvl="6"],
that's rather stretching it; it might as well be the same colour as the
sidebar background. I've investigated making the tab colour lighter by
modifying its HSL value, but the results aren't as consistent.