Rework Sidebery auto-collapse CSS
Following the rework in cb48760, the width of the sidebar can no longer
be used as the condition for auto-collapse. Thus, the condition is
adjusted to when the sidebar is hovered. However, this new condition is
more fragile, as it's invalidated as soon as the cursor moves out of the
sidebar area.

To address this, I've added some transition-delay to ensure the
invalidation doesn't happen too often. This prevents the elements from
disappearing/appearing too frequently. Unfortunately, I can't find a way
to delay setting `tab-indent: 0`, so Sidebery will always indent/
deindent when the cursor hovers over it.