Remove top border when tabs toolbar is collapsed
A 1px border will persists above the navigation bar when the tabs
toolbar is collapsed. This was caused by an upstream commit in [bug 1889463](https://bugzilla.mozilla.org/show_bug.cgi?id=1889463).
While I'm not inherently against the appearance of the top border, it
unknowingly hurts usability - Fitts's Law. The top border prevents
the user from clicking the Back, Forward and Reload button if their
cursors are at the top left corner of the screen. This is because the
border property takes up space and alters the layout slightly.