Delay collapsing tabs tree when sidebar is unhovered briefly
Previously, the tabs tree would immediately collapse/unindent as soon as
the sidebar was unhovered. This constant repainting whenever the cursor
moves in or out of the sidebar could reduce perceived performance,
which I've been trying to optimize since a5aef47. Now that it's possible
to animate custom properties, I can delay setting `--tabs-indent` to 0.
[1] The downside is that I haven't figure out how to transition a custom
property, so there's no animation for collapsing/uncollapsing the tabs.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/@property