Rework hide tabs toolbar CSS
Due to the ongoing work on vertical tabs and sidebar revamp in Nightly,
major changes in the CSS source files are expected. As a result,
these userChrome CSS styles need to be updated to accommodate
the upstream changes.

- Migrate to hide_tabs_toolbar_v2.css
  Source: https://github.com/MrOtherGuy/firefox-csshacks/blob/master/chrome/hide_tabs_toolbar_v2.css
  Relevant issues: MrOtherGuy/firefox-csshacks#425,
                   MrOtherGuy/firefox-csshacks#413,
                   MrOtherGuy/firefox-csshacks#426
- Remove hide_tabs_toolbar.css and windows_controls_placeholder.css

Note: `hide_tabs_toolbar_v2.css` is still a prototype, thus not all
      CSS styles from v1 are implemented, notably;
      - Inline menubar with fade out
      - A thin strip is present above the navbar
      - No smooth transition when collapsing the tabs toolbar