Add info about a related PDF-downloading pref
From what I can tell, `browser.download.force_save_internally_handled_attachments`
is mutually exclusive with `browser.download.open_pdf_attachments_inline`.

According to a Mozilla Connect comment[1], the former forces Firefox to
only save PDF (when `Content-Disposition` is set to `download`), and
doesn't open it afterwards (or asks where to save if configured in
`about:preferences`). The catch is that Firefox treats PDF as "native
file types" and defaults to opening them in Firefox. This leads to a
bizarre behaviour where Firefox downloads the PDF (disregarding the
"Ask where to save files" pref, mind you), then erratically opens it.
This is disgustingly untuitive, and it's still the default.

The latter, however, always opens PDF in Firefox and never downloads it.
It doesn't respect the `Content-Disposition` flag, but at least it lets
me decide which PDFs to download.

You can use this site to test the difference:
https://emk.name/test/bug1790641.html

[1]: https://connect.mozilla.org/t5/discussions/stop-saving-pdfs-to-downloads-when-viewing-in-firefox/m-p/78399/highlight/true#M29939